requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2023.2
sqlparse==0.4.3
